changes on login and bookings using swal

+++ src/Components/Bookings/Bookings.tsx
@@ -11,6 +11,7 @@ import { NavLink } from "react-router-dom";
 import { useCustomDispatch, useCustomSelector } from '../../hooks/redux/index';
 import { FilterButton, Input, SearchIcon, Select, WrapperButton, WrapperInput, Option } from '../StyledFilterButtons';
 import { colors } from '../theme';
+import Swal from "sweetalert2";
 
 
 export const Bookings = () => {
@@ -24,8 +25,28 @@ useEffect(() => {
 }, [dispatch])
 
 const onDeleteBooking = (bookingId: string) => {
-  dispatch(deleteBooking(bookingId))
-}
+  // Muestra la alerta de confirmación antes de eliminar
+  Swal.fire({
+    title: 'Are you sure?',
+    text: "You won't be able to revert this!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete it!',
+  }).then((result) => {
+    if (result.isConfirmed) {
+      // Si el usuario confirma, ejecuta la eliminación
+      dispatch(deleteBooking(bookingId));
+      // Muestra la alerta de éxito después de eliminar
+      Swal.fire({
+        title: 'Deleted!',
+        text: 'The booking has been deleted.',
+        icon: 'success',
+      });
+    }
+  });
+};
 
 interface PopUpStates {
   [bookingId: string]: boolean;
